Clean migrations, fix big with calling generator of background profile picture

+++ AHC_app/medical_notes/type_measurement_notes/models.py
@@ -1,17 +1,16 @@
-from django.db import models
-
 from animals.models import Animal
+from django.db import models
 from medical_notes.models import MedicalRecord
 
 
 class BiometricHeightRecords(models.Model):
-    height = models.IntegerField(default=0)
+    height = models.DecimalField(default=0, max_digits=8, decimal_places=3)
     # always in grams, set validation to int values (if is float, ask if save as integer grams)
     height_unit_to_present = models.CharField(max_length=3, default="mm", blank=False)
 
 
 class BiometricWeightRecords(models.Model):
-    weight = models.IntegerField(default=0)
+    weight = models.DecimalField(default=0, max_digits=8, decimal_places=3)
     # always in mm, set validation to int values (if is float, ask if save as integer grams)
     weight_unit_to_present = models.CharField(max_length=3, default="g", blank=False)
 
@@ -24,9 +23,17 @@ class BiometricCustomRecords(models.Model):
 
 class BiometricRecord(models.Model):
     animal = models.ForeignKey(Animal, on_delete=models.CASCADE)
-    related_note = models.ForeignKey(MedicalRecord, on_delete=models.SET_NULL, blank=True, null=True)
+    related_note = models.ForeignKey(
+        MedicalRecord, on_delete=models.SET_NULL, blank=True, null=True
+    )
     date_updated = models.DateTimeField(auto_now_add=True, editable=True)
 
-    weight_biometric_record = models.OneToOneField(BiometricWeightRecords, on_delete=models.CASCADE, blank=True, null=True)
-    height_biometric_record = models.OneToOneField(BiometricHeightRecords, on_delete=models.CASCADE, blank=True, null=True)
-    custom_biometric_record = models.OneToOneField(BiometricCustomRecords, on_delete=models.CASCADE, blank=True, null=True)
+    weight_biometric_record = models.OneToOneField(
+        BiometricWeightRecords, on_delete=models.CASCADE, blank=True, null=True
+    )
+    height_biometric_record = models.OneToOneField(
+        BiometricHeightRecords, on_delete=models.CASCADE, blank=True, null=True
+    )
+    custom_biometric_record = models.OneToOneField(
+        BiometricCustomRecords, on_delete=models.CASCADE, blank=True, null=True
+    )
